requirements.txt
langchain==1.2.6
langchain-community==0.4.1
langchain-text-splitters==1.1.0
langchain-core==1.2.7
langchain-groq==1.1.1
langchain-google-genai==4.2.0
langchain-chroma==1.1.2
chromadb==1.4.1
beautifulsoup4==4.11.1
python-dotenv==1.0.1